+++ game_patch/graphics/gr.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "../rf/bmpman.h"
+#include "../rf/gr/gr.h"
 
 void gr_apply_patch();
 void gr_render_fps_counter();
@@ -10,6 +11,7 @@ bool gr_render_to_texture(int bm_handle);
 void gr_render_to_back_buffer();
 void gr_delete_texture(int bm_handle);
 bool gr_is_texture_format_supported(rf::BmFormat format);
+void gr_bitmap_stretched_float(int bitmap_handle, float x, float y, float w, float h, float sx, float sy, float sw, float sh, bool flip_x, bool flip_y, rf::GrMode mode);
 
 template<typename F>
 void gr_font_run_with_default(int font_id, F fun)