+++ game_patch/misc/player_fpgun.cpp
@@ -1,14 +1,16 @@
+#include <patch_common/CodeInjection.h>
+#include <patch_common/FunHook.h>
+#include <patch_common/CallHook.h>
+#include <patch_common/AsmWriter.h>
+#include <common/config/BuildConfig.h>
 #include "../rf/player.h"
 #include "../rf/sound/sound.h"
 #include "../rf/vmesh.h"
 #include "../rf/weapon.h"
 #include "../rf/entity.h"
-#include <patch_common/CodeInjection.h>
-#include <patch_common/FunHook.h>
-#include <patch_common/AsmWriter.h>
-#include <common/config/BuildConfig.h>
 #include "../graphics/gr.h"
 #include "../main/main.h"
+#include "../os/console.h"
 
 static std::vector<int> g_fpgun_sounds;
 
@@ -100,6 +102,26 @@ CodeInjection after_game_render_to_dynamic_textures{
     },
 };
 
+CallHook<void(rf::Matrix3&, rf::Vector3&, float, bool, bool)> player_fpgun_render_gr_setup_3d_hook{
+    0x004AB411,
+    [](rf::Matrix3& viewer_orient, rf::Vector3& viewer_pos, float horizontal_fov, bool zbuffer_flag, bool z_scale) {
+        horizontal_fov *= g_game_config.fpgun_fov_scale;
+        player_fpgun_render_gr_setup_3d_hook
+            .call_target(viewer_orient, viewer_pos, horizontal_fov, zbuffer_flag, z_scale);
+    },
+};
+
+ConsoleCommand2 fpgun_fov_scale_cmd{
+    "fpgun_fov_scale",
+    [](std::optional<float> scale_opt) {
+        if (scale_opt) {
+            g_game_config.fpgun_fov_scale = std::clamp(scale_opt.value(), 0.2f, 2.0f);
+            g_game_config.save();
+        }
+        rf::console_printf("Fpgun FOV scale: %.4f", g_game_config.fpgun_fov_scale.value());
+    },
+};
+
 void player_fpgun_do_patch()
 {
 #if SPECTATE_MODE_SHOW_WEAPON
@@ -157,4 +179,8 @@ void player_fpgun_do_patch()
 
     // Render rocket launcher scanner image every frame
     // addr_as_ref<bool>(0x5A1020) = 0;
+
+    // Allow customizing fpgun fov
+    player_fpgun_render_gr_setup_3d_hook.install();
+    fpgun_fov_scale_cmd.register_cmd();
 }