+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ Version 1.6.1 (not released yet)
 - Remove UID limit (50k) in editor
 - Fix shape and script name being overwritten if editing properties of multiple triggers that have those fields set
   differently
+- Do not verify mod vpp checksum if `Require Client Mod` option is disabled in server config
 
 Version 1.6.1-beta (released 2021-01-29)
 ----------------------------------------